+++ SIL.Core/Xml/XmlUtils.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Linq;
@@ -7,6 +8,7 @@ using System.Text;
 using System.Xml;
 using System.Xml.Linq;
 using System.Xml.XPath;
+using JetBrains.Annotations;
 using SIL.Linq;
 
 namespace SIL.Xml
@@ -22,6 +24,7 @@ namespace SIL.Xml
 		/// <param name="node">The XmlNode to look in.</param>
 		/// <param name="attrName">The optional attribute to find.</param>
 		/// <returns></returns>
+		[PublicAPI]
 		public static bool GetBooleanAttributeValue(XmlNode node, string attrName)
 		{
 			return GetBooleanAttributeValue(GetOptionalAttributeValue(node, attrName));
@@ -39,9 +42,9 @@ namespace SIL.Xml
 		}
 
 		/// <summary>
-		/// Given bytes that represent an xml element, return the values of requested attributes (if they exist).
+		/// Given bytes that represent an XML element, return the values of requested attributes (if they exist).
 		/// </summary>
-		/// <param name="data">Data that is expected to an xml element.</param>
+		/// <param name="data">Data that is expected to an XML element.</param>
 		/// <param name="attributes">A set of attributes, the values of which are to be returned.</param>
 		/// <returns>A dictionary </returns>
 		public static Dictionary<string, string> GetAttributes(byte[] data, HashSet<string> attributes)
@@ -63,9 +66,9 @@ namespace SIL.Xml
 		}
 
 		/// <summary>
-		/// Given a string that represents an xml element, return the values of requested attributes (if they exist).
+		/// Given a string that represents an XML element, return the values of requested attributes (if they exist).
 		/// </summary>
-		/// <param name="data">Data that is expected to an xml element.</param>
+		/// <param name="data">Data that is expected to an XML element.</param>
 		/// <param name="attributes">A set of attributes, the values of which are to be returned.</param>
 		/// <returns>A dictionary </returns>
 		public static Dictionary<string, string> GetAttributes(string data, HashSet<string> attributes)
@@ -92,6 +95,7 @@ namespace SIL.Xml
 		/// <param name="node">The XmlNode to look in.</param>
 		/// <param name="attrName">The optional attribute to find.</param>
 		/// <returns></returns>
+		[PublicAPI]
 		public static bool GetBooleanAttributeValue(XPathNavigator node, string attrName)
 		{
 			return GetBooleanAttributeValue(GetOptionalAttributeValue(node, attrName));
@@ -107,18 +111,19 @@ namespace SIL.Xml
 		}
 
 		/// <summary>
-		/// Returns a integer obtained from the (mandatory) attribute named.
+		/// Returns an integer obtained from the (mandatory) attribute named.
 		/// </summary>
 		/// <param name="node">The XmlNode to look in.</param>
 		/// <param name="attrName">The mandatory attribute to find.</param>
 		/// <returns>The value, or 0 if attr is missing.</returns>
+		[PublicAPI]
 		public static int GetMandatoryIntegerAttributeValue(XmlNode node, string attrName)
 		{
-			return Int32.Parse(GetMandatoryAttributeValue(node, attrName));
+			return int.Parse(GetMandatoryAttributeValue(node, attrName));
 		}
 
 		/// <summary>
-		/// Returns a integer obtained from the (mandatory) attribute named.
+		/// Returns an integer obtained from the (mandatory) attribute named.
 		/// </summary>
 		/// <param name="element">The XmlNode to look in.</param>
 		/// <param name="attrName">The mandatory attribute to find.</param>
@@ -137,6 +142,7 @@ namespace SIL.Xml
 		/// <param name="attrName"></param>
 		/// <param name="defaultVal"></param>
 		/// <returns></returns>
+		[PublicAPI]
 		public static int GetOptionalIntegerValue(XmlNode node, string attrName, int defaultVal)
 		{
 			string val = GetOptionalAttributeValue(node, attrName);
@@ -144,7 +150,7 @@ namespace SIL.Xml
 			{
 				return defaultVal;
 			}
-			return Int32.Parse(val);
+			return int.Parse(val);
 		}
 
 		/// <summary>
@@ -163,6 +169,7 @@ namespace SIL.Xml
 		/// <param name="node"></param>
 		/// <param name="attrName"></param>
 		/// <returns></returns>
+		[PublicAPI]
 		public static int[] GetMandatoryIntegerListAttributeValue(XmlNode node, string attrName)
 		{
 			string input = GetMandatoryAttributeValue(node, attrName);
@@ -170,7 +177,7 @@ namespace SIL.Xml
 			int[] result = new int[vals.Length];
 			for (int i = 0;i < vals.Length;i++)
 			{
-				result[i] = Int32.Parse(vals[i]);
+				result[i] = int.Parse(vals[i]);
 			}
 			return result;
 		}
@@ -198,6 +205,7 @@ namespace SIL.Xml
 		/// <param name="attrName"></param>
 		/// <returns></returns>
 		[CLSCompliant(false)]
+		[PublicAPI]
 		public static uint[] GetMandatoryUIntegerListAttributeValue(XmlNode node, string attrName)
 		{
 			var input = GetMandatoryAttributeValue(node, attrName);
@@ -240,6 +248,7 @@ namespace SIL.Xml
 		/// </summary>
 		/// <param name="vals"></param>
 		/// <returns></returns>
+		[PublicAPI]
 		public static string MakeIntegerListValue(int[] vals)
 		{
 			return MakeStringFromEnum(vals, vals.Length);
@@ -273,6 +282,7 @@ namespace SIL.Xml
 		/// <param name="attrName">The attribute to find.</param>
 		/// <param name="defaultValue"></param>
 		/// <returns>The value of the attribute, or the default value, if the attribute dismissing</returns>
+		[PublicAPI]
 		public static bool GetOptionalBooleanAttributeValue(XmlNode node, string attrName,
 			bool defaultValue)
 		{
@@ -336,7 +346,7 @@ namespace SIL.Xml
 		public static string GetOptionalAttributeValue(XmlNode node, string attrName,
 			string defaultString)
 		{
-			if (node != null && node.Attributes != null)
+			if (node is { Attributes: { } })
 			{
 				XmlAttribute xa = node.Attributes[attrName];
 				if (xa != null)
@@ -358,9 +368,7 @@ namespace SIL.Xml
 			string defaultString)
 		{
 			if (element == null || !element.Attributes().Any())
-			{
 				return defaultString;
-			}
 			var attribute = element.Attribute(attrName);
 			return attribute != null ? attribute.Value : defaultString;
 		}
@@ -375,13 +383,11 @@ namespace SIL.Xml
 		public static string GetOptionalAttributeValue(XPathNavigator node, string attrName,
 			string defaultString)
 		{
-			if (node != null && node.HasAttributes)
+			if (node is { HasAttributes: true })
 			{
 				string s = node.GetAttribute(attrName, string.Empty);
 				if (!string.IsNullOrEmpty(s))
-				{
 					return s;
-				}
 			}
 			return defaultString;
 		}
@@ -395,20 +401,16 @@ namespace SIL.Xml
 		public static XmlNode FindNode(XmlNode node, string name)
 		{
 			if (node.Name == name)
-			{
 				return node;
-			}
+
 			foreach (XmlNode childNode in node.ChildNodes)
 			{
 				if (childNode.Name == name)
-				{
 					return childNode;
-				}
+
 				XmlNode n = FindNode(childNode, name);
 				if (n != null)
-				{
 					return n;
-				}
 			}
 			return null;
 		}
@@ -489,6 +491,7 @@ namespace SIL.Xml
 			return retval;
 		}
 
+		[PublicAPI]
 		public static string GetMandatoryAttributeValue(XPathNavigator node, string attrName)
 		{
 			string retval = GetOptionalAttributeValue(node, attrName, null);
@@ -499,7 +502,7 @@ namespace SIL.Xml
 		}
 
 		/// <summary>
-		/// Append an child node with the specified name and value to <paramref name="node"/>.
+		/// Append a child node with the specified name and value to <paramref name="node"/>.
 		/// </summary>
 		/// <param name="node"></param>
 		/// <param name="elementName"></param>
@@ -556,35 +559,28 @@ namespace SIL.Xml
 		public static bool NodesMatch(XmlNode node1, XmlNode node2)
 		{
 			if (node1 == null && node2 == null)
-			{
 				return true;
-			}
+
 			if (node1 == null || node2 == null)
-			{
 				return false;
-			}
+
 			if (node1.Name != node2.Name)
-			{
 				return false;
-			}
+
 			if (node1.InnerText != node2.InnerText)
-			{
 				return false;
-			}
+
 			if (node1.Attributes == null && node2.Attributes != null)
-			{
 				return false;
-			}
-			if (node1.Attributes != null && node2.Attributes == null)
-			{
-				return false;
-			}
+
 			if (node1.Attributes != null)
 			{
-				if (node1.Attributes.Count != node2.Attributes.Count)
-				{
+				if (node2.Attributes == null)
 					return false;
-				}
+
+				if (node1.Attributes.Count != node2.Attributes.Count)
+					return false;
+
 				for (int i = 0;i < node1.Attributes.Count;i++)
 				{
 					XmlAttribute xa1 = node1.Attributes[i];
@@ -595,53 +591,41 @@ namespace SIL.Xml
 					}
 				}
 			}
-			if (node1.ChildNodes == null && node2.ChildNodes != null)
+
+			int iChild1 = 0; // index node1.ChildNodes
+			int iChild2 = 0; // index node2.ChildNodes
+			while (iChild1 < node1.ChildNodes.Count && iChild2 < node1.ChildNodes.Count)
 			{
-				return false;
-			}
-			if (node1.ChildNodes != null && node2.ChildNodes == null)
-			{
-				return false;
-			}
-			if (node1.ChildNodes != null)
-			{
-				int ichild1 = 0; // index node1.ChildNodes
-				int ichild2 = 0; // index node2.ChildNodes
-				while (ichild1 < node1.ChildNodes.Count && ichild2 < node1.ChildNodes.Count)
+				XmlNode child1 = node1.ChildNodes[iChild1];
+
+				// Note that we must defer doing the 'continue' until after we have checked to see if both children are comments
+				// If we continue immediately and the last node of both elements is a comment, the second node will not have
+				// iChild2 incremented and the final test will fail.
+				bool foundComment = false;
+
+				if (child1 is XmlComment)
 				{
-					XmlNode child1 = node1.ChildNodes[ichild1];
-
-					// Note that we must defer doing the 'continue' until after we have checked to see if both children are comments
-					// If we continue immediately and the last node of both elements is a comment, the second node will not have
-					// ichild2 incremented and the final test will fail.
-					bool foundComment = false;
-
-					if (child1 is XmlComment)
-					{
-						ichild1++;
-						foundComment = true;
-					}
-					XmlNode child2 = node2.ChildNodes[ichild2];
-					if (child2 is XmlComment)
-					{
-						ichild2++;
-						foundComment = true;
-					}
-
-					if (foundComment)
-					{
-						continue;
-					}
-
-					if (!NodesMatch(child1, child2))
-					{
-						return false;
-					}
-					ichild1++;
-					ichild2++;
+					iChild1++;
+					foundComment = true;
 				}
+				XmlNode child2 = node2.ChildNodes[iChild2];
+				if (child2 is XmlComment)
+				{
+					iChild2++;
+					foundComment = true;
+				}
+
+				if (foundComment)
+					continue;
+
+				if (!NodesMatch(child1, child2))
+					return false;
+
+				iChild1++;
+				iChild2++;
+				
 				// If we finished both lists we got a match.
-				return ichild1 == node1.ChildNodes.Count && ichild2 == node2.ChildNodes.Count;
+				return iChild1 == node1.ChildNodes.Count && iChild2 == node2.ChildNodes.Count;
 			}
 			// both lists are null
 			return true;
@@ -683,20 +667,20 @@ namespace SIL.Xml
 				return false;
 			if (element1.HasElements)
 			{
-				int ichild1 = 0; // index node1.Elements()
-				int ichild2 = 0; // index node2.Elements()
-				while (ichild1 < element1.Elements().Count() && ichild2 < element1.Elements().Count())
+				int iChild1 = 0; // index node1.Elements()
+				int iChild2 = 0; // index node2.Elements()
+				while (iChild1 < element1.Elements().Count() && iChild2 < element1.Elements().Count())
 				{
-					var child1 = element1.Elements().ToList()[ichild1];
-					var child2 = element2.Elements().ToList()[ichild2];
+					var child1 = element1.Elements().ToList()[iChild1];
+					var child2 = element2.Elements().ToList()[iChild2];
 
 					if (!NodesMatch(child1, child2))
 						return false;
-					ichild1++;
-					ichild2++;
+					iChild1++;
+					iChild2++;
 				}
 				// If we finished both lists we got a match.
-				return (ichild1 == element1.Elements().Count()) && (ichild2 == element2.Elements().Count());
+				return iChild1 == element1.Elements().Count() && iChild2 == element2.Elements().Count();
 			}
 
 			// both lists are null
@@ -707,7 +691,7 @@ namespace SIL.Xml
 		/// Return the first child of the node that is not a comment (or null).
 		/// </summary>
 		/// <param name="node"></param>
-		/// <returns></returns>
+		[PublicAPI]
 		public static XmlNode GetFirstNonCommentChild(XmlNode node)
 		{
 			if (node == null)
@@ -729,9 +713,10 @@ namespace SIL.Xml
 		/// </summary>
 		/// <param name="element"></param>
 		/// <returns></returns>
+		[PublicAPI]
 		public static XElement GetFirstNonCommentChild(XElement element)
 		{
-			return (element == null) ? null : element.Elements().FirstOrDefault();
+			return element?.Elements().FirstOrDefault();
 		}
 
 		/// <summary>
@@ -753,8 +738,9 @@ namespace SIL.Xml
 		}
 
 		/// <summary>
-		/// Convert a possibly multiparagraph string to a form that is safe to store both in an XML file.
+		/// Convert a possibly multi-paragraph string to a form that is safe to store both in an XML file.
 		/// </summary>
+		[PublicAPI]
 		public static string ConvertMultiparagraphToSafeXml(string sInput)
 		{
 			var sOutput = sInput;
@@ -774,6 +760,7 @@ namespace SIL.Xml
 		/// </summary>
 		/// <param name="sInput"></param>
 		/// <returns></returns>
+		[PublicAPI]
 		public static string MakeSafeXmlAttribute(string sInput)
 		{
 			string sOutput = sInput;
@@ -793,11 +780,12 @@ namespace SIL.Xml
 		/// Convert an encoded attribute string into plain text.
 		/// </summary>
 		/// <param name="sInput"></param>
-		/// <returns></returns>
+		/// <returns>The plain text representation, with all the XML codes replaced by their normal
+		/// (single-character) textual representations</returns>
 		public static string DecodeXmlAttribute(string sInput)
 		{
 			string sOutput = sInput;
-			if (!String.IsNullOrEmpty(sOutput) && sOutput.Contains("&"))
+			if (!string.IsNullOrEmpty(sOutput) && sOutput.Contains("&"))
 			{
 				sOutput = sOutput.Replace("&gt;", ">");
 				sOutput = sOutput.Replace("&lt;", "<");
@@ -805,6 +793,7 @@ namespace SIL.Xml
 				sOutput = sOutput.Replace("&quot;", "\"");
 				sOutput = sOutput.Replace("&amp;", "&");
 			}
+
 			for (int idx = sOutput.IndexOf("&#"); idx >= 0; idx = sOutput.IndexOf("&#"))
 			{
 				int idxEnd = sOutput.IndexOf(';', idx);
@@ -813,19 +802,17 @@ namespace SIL.Xml
 				string sOrig = sOutput.Substring(idx, (idxEnd - idx) + 1);
 				string sNum = sOutput.Substring(idx + 2, idxEnd - (idx + 2));
 				string sReplace = null;
-				int chNum = 0;
 				if (sNum[0] == 'x' || sNum[0] == 'X')
 				{
-					if (Int32.TryParse(sNum.Substring(1), NumberStyles.HexNumber, NumberFormatInfo.InvariantInfo, out chNum))
-						sReplace = Char.ConvertFromUtf32(chNum);
+					if (int.TryParse(sNum.Substring(1), NumberStyles.HexNumber, NumberFormatInfo.InvariantInfo, out var chNum))
+						sReplace = char.ConvertFromUtf32(chNum);
 				}
 				else
 				{
-					if (Int32.TryParse(sNum, out chNum))
-						sReplace = Char.ConvertFromUtf32(chNum);
+					if (int.TryParse(sNum, out var chNum))
+						sReplace = char.ConvertFromUtf32(chNum);
 				}
-				if (sReplace == null)
-					sReplace = sNum;
+				sReplace ??= sNum;
 				sOutput = sOutput.Replace(sOrig, sReplace);
 			}
 			return sOutput;
@@ -837,13 +824,13 @@ namespace SIL.Xml
 		/// <param name="xml"></param>
 		/// <returns></returns>
 		public static string GetIndentedXml(string xml)
-	  {
-		 string outXml = string.Empty;
-		 using(MemoryStream ms = new MemoryStream())
-		 // Create a XMLTextWriter that will send its output to a memory stream (file)
-		 using (XmlTextWriter xtw = new XmlTextWriter(ms, Encoding.Unicode))
-		 {
-			 XmlDocument doc = new XmlDocument();
+		{
+			string outXml = string.Empty;
+			using(MemoryStream ms = new MemoryStream())
+			// Create a XMLTextWriter that will send its output to a memory stream (file)
+			using (XmlTextWriter xtw = new XmlTextWriter(ms, Encoding.Unicode))
+			{
+				XmlDocument doc = new XmlDocument();
 
 //             try
 //             {
@@ -851,7 +838,7 @@ namespace SIL.Xml
 				 // of the XML Document Object Model (DOM)
 				 doc.LoadXml(xml);
 
-				 // Set the formatting property of the XML Text Writer to indented
+				 // Set the formatting property of the XML Text Writer to be indented
 				 // the text writer is where the indenting will be performed
 				 xtw.Formatting = Formatting.Indented;
 
@@ -873,11 +860,11 @@ namespace SIL.Xml
 //             {
 //                 return ex.Message;
 //             }
-		 }
-	  }
+			}
+		}
 
-		//todo: what's the diff between this one and the next?
-		static public XmlElement GetOrCreateElementPredicate(XmlDocument dom, XmlElement parent, string predicate, string name)
+		[PublicAPI]
+		public static XmlElement GetOrCreateElementPredicate(XmlDocument dom, XmlElement parent, string predicate, string name)
 		{
 			XmlElement element = (XmlElement)parent.SelectSingleNodeHonoringDefaultNS("/" + predicate);
 			if (element == null)
@@ -888,7 +875,7 @@ namespace SIL.Xml
 			return element;
 		}
 
-		static public XmlElement GetOrCreateElement(XmlDocument dom, string parentPath, string name)
+		public static XmlElement GetOrCreateElement(XmlDocument dom, string parentPath, string name)
 		{
 			XmlElement element = (XmlElement)dom.SelectSingleNodeHonoringDefaultNS(parentPath + "/" + name);
 			if (element == null)
@@ -910,29 +897,27 @@ namespace SIL.Xml
 			}
 			catch (NullReferenceException)
 			{
-				throw new XmlFormatException(string.Format("Expected a {0} attribute on {1}.", attr, form.OuterXml));
+				throw new XmlFormatException($"Expected a {attr} attribute on {form.OuterXml}.");
 			}
 		}
 
 		public static string GetOptionalAttributeString(XmlNode xmlNode, string attributeName)
 		{
-			XmlAttribute attr = xmlNode.Attributes[attributeName];
-			if (attr == null)
-				return null;
-			return attr.Value;
+			return xmlNode.Attributes[attributeName]?.Value;
 		}
 
+		[PublicAPI]
 		public static XmlNode GetDocumentNodeFromRawXml(string outerXml, XmlNode nodeMaker)
 		{
 			if (string.IsNullOrEmpty(outerXml))
-			{
 				throw new ArgumentException();
-			}
-			XmlDocument doc = nodeMaker as XmlDocument;
-			if (doc == null)
+
+			if (!(nodeMaker is XmlDocument doc))
 			{
 				doc = nodeMaker.OwnerDocument;
+				Debug.Assert(doc != null);
 			}
+
 			using (StringReader sr = new StringReader(outerXml))
 			{
 				using (XmlReader r = XmlReader.Create(sr))
@@ -943,15 +928,16 @@ namespace SIL.Xml
 			}
 		}
 
+		[PublicAPI]
 		public static string GetXmlForShowingInHtml(string xml)
 		{
-			var s = XmlUtils.GetIndentedXml(xml).Replace("<", "&lt;");
+			var s = GetIndentedXml(xml).Replace("<", "&lt;");
 			s = s.Replace("\r\n", "<br/>");
 			s = s.Replace("  ", "&nbsp;&nbsp;");
 			return s;
 		}
 
-
+		[PublicAPI]
 		public static string GetTitleOfHtml(XmlDocument dom, string defaultIfMissing)
 		{
 			var title = dom.SelectSingleNode("//head/title");
@@ -1040,8 +1026,7 @@ namespace SIL.Xml
 				writer.WriteString("");
 			foreach (var child in element.Nodes())
 			{
-				var xElement = child as XElement;
-				if (xElement != null)
+				if (child is XElement xElement)
 					WriteElementTo(writer, xElement, suppressIndentingChildren, preserveNamespaces);
 				else
 					child.WriteTo(writer); // defaults are fine for everything else.
@@ -1066,15 +1051,15 @@ namespace SIL.Xml
 				int code;
 				try
 				{
-					code = Char.ConvertToUtf32(s, i);
+					code = char.ConvertToUtf32(s, i);
 				}
 				catch (ArgumentException)
 				{
 					continue;
 				}
 				if (IsLegalXmlChar(code))
-					buffer.Append(Char.ConvertFromUtf32(code));
-				if (Char.IsSurrogatePair(s, i))
+					buffer.Append(char.ConvertFromUtf32(code));
+				if (char.IsSurrogatePair(s, i))
 					i++;
 			}
 
@@ -1139,13 +1124,11 @@ namespace SIL.Xml
 						fSuccessfulVisit = true;
 				}
 			}
-			if (input.ChildNodes != null) // not sure whether this can happen.
+
+			foreach (XmlNode child in input.ChildNodes)
 			{
-				foreach (XmlNode child in input.ChildNodes)
-				{
-					if (VisitAttributes(child, visitor))
-						fSuccessfulVisit = true;
-				}
+				if (VisitAttributes(child, visitor))
+					fSuccessfulVisit = true;
 			}
 			return fSuccessfulVisit;
 		}
@@ -1187,23 +1170,39 @@ namespace SIL.Xml
 		bool Visit(XAttribute xa);
 	}
 
+	/// <summary>
+	/// "Visitor" that replaces a substring in the value of an attribute.
+	/// </summary>
 	public class ReplaceSubstringInAttr : IAttributeVisitor
 	{
 		readonly string _pattern;
 		readonly string _replacement;
+
+		/// <summary>
+		/// Sets up an object that can be used to replace a particular substring in the value of an
+		/// attribute.
+		/// </summary>
+		/// <param name="pattern">The text of the substring to search for (using case-sensitive,
+		/// ordinal comparison)</param>
+		/// <param name="replacement">The replacement text</param>
 		public ReplaceSubstringInAttr(string pattern, string replacement)
 		{
 			_pattern = pattern;
 			_replacement = replacement;
 		}
+
+		/// <summary>
+		/// Actually performs the replacement in the given attribute.
+		/// </summary>
+		/// <returns>Whether the pattern was found and the replacement made</returns>
 		public virtual bool Visit(XAttribute xa)
 		{
 			string old = xa.Value;
-			int index = old.IndexOf(_pattern);
+			int index = old.IndexOf(_pattern, StringComparison.Ordinal);
 			if (index < 0)
 				return false;
 			xa.Value = old.Replace(_pattern, _replacement);
-			return false;
+			return true;
 		}
 	}
 }