+++ SIL.Lift/Options/Option.cs
@@ -12,23 +12,19 @@ namespace SIL.Lift.Options
 		private MultiText _abbreviation;
 		private MultiText _description;
 		private string _humanReadableKey;
-		private MultiText _name;
 
 		public Option(): this(string.Empty, new MultiText()) {}
 
 		public Option(string humanReadableKey, MultiText name) //, Guid guid)
 		{
 			_humanReadableKey = humanReadableKey;
-			_name = name;
+			Name = name;
 			//SearchKeys = new List<string>();
 		}
 
 		#region IChoice Members
 
-		public string Label
-		{
-			get { return _name.GetFirstAlternative(); }
-		}
+		public string Label => Name.GetFirstAlternative();
 
 		#endregion
 
@@ -38,14 +34,9 @@ namespace SIL.Lift.Options
 			get
 			{
 				if (String.IsNullOrEmpty(_humanReadableKey))
-				{
 					return GetDefaultKey(); //don't actually save it yet
-				}
 
-				else
-				{
-					return _humanReadableKey;
-				}
+				return _humanReadableKey;
 			}
 
 			set
@@ -54,8 +45,9 @@ namespace SIL.Lift.Options
 				{
 					_humanReadableKey = GetDefaultKey();
 				}
-						//the idea here is, we're delaying setting the key in concrete for as long as possible
-						//this allows the ui to continue to auto-create the key during a ui session.
+
+				// The idea here is, we're delaying setting the key in concrete for as long as
+				// possible to allow the UI to continue to auto-create the key during a UI session.
 				else if (value != GetDefaultKey())
 				{
 					_humanReadableKey = value.Trim();
@@ -66,26 +58,15 @@ namespace SIL.Lift.Options
 		//        [ReflectorProperty("name", typeof (MultiTextSerializorFactory),
 		//            Required = true)]
 		[XmlElement("name")]
-		public MultiText Name
-		{
-			get { return _name; }
-			set { _name = value; }
-		}
+		public MultiText Name { get; set; }
 
 		//        [ReflectorProperty("abbreviation", typeof (MultiTextSerializorFactory),
 		//            Required = false)]
 		[XmlElement("abbreviation")]
 		public MultiText Abbreviation
 		{
-			get
-			{
-				if (_abbreviation == null)
-				{
-					return Name;
-				}
-				return _abbreviation;
-			}
-			set { _abbreviation = value; }
+			get => _abbreviation ?? Name;
+			set => _abbreviation = value;
 		}
 
 		//
@@ -97,12 +78,11 @@ namespace SIL.Lift.Options
 			get
 			{
 				if (_description == null)
-				{
 					_description = new MultiText();
-				}
+
 				return _description;
 			}
-			set { _description = value; }
+			set => _description = value;
 		}
 
 		[XmlElement("searchKeys")]
@@ -135,7 +115,7 @@ namespace SIL.Lift.Options
 
 		public override string ToString()
 		{
-			return _name.GetFirstAlternative();
+			return Name.GetFirstAlternative();
 		}
 
 		public object GetDisplayProxy(string writingSystemId)
@@ -181,10 +161,10 @@ namespace SIL.Lift.Options
 		public IList<string> GetSearchKeys(string writingSystemId)
 		{
 			var keys = SearchKeys;
-			if(keys==null)
+			if (keys == null)
 				return new List<string>();
 			var alt = SearchKeys.GetExactAlternative(writingSystemId);
-			if(alt==null)
+			if (alt == null)
 				return new List<string>();
 
 			return alt.SplitTrimmed(',');