+++ SIL.Windows.Forms/FileDialogExtender/FileDialogControlBase.cs
@@ -1,5 +1,5 @@
 //  Copyright (c) 2006, Gustavo Franco
-//  Copyright © Decebal Mihailescu 2007-2010
+//  Copyright Â© Decebal Mihailescu 2007-2010
 
 //  Email:  gustavo_franco@hotmail.com
 //  All rights reserved.
@@ -60,18 +60,14 @@ namespace SIL.Windows.Forms.FileDialogExtender
 		#endregion
 
 		#region Variables Declaration
-		System.Windows.Forms.FileDialog _MSdialog;
+		FileDialog _MSdialog;
 		NativeWindow _dlgWrapper;
 		private AddonWindowLocation _StartLocation = AddonWindowLocation.Right;
-		private FolderViewMode _DefaultViewMode = FolderViewMode.Default;
 		IntPtr _hFileDialogHandle = IntPtr.Zero;
-		FileDialogType _FileDlgType;
 		string _InitialDirectory = string.Empty;
 		string _Filter = "All files (*.*)|*.*";
 		string _DefaultExt = "jpg";
 		string _FileName = string.Empty;
-		string _Caption = "Save";
-		string _OKCaption = "&Open";
 		int _FilterIndex = 1;
 		bool _AddExtension = true;
 		bool _CheckFileExists = true;
@@ -98,14 +94,14 @@ namespace SIL.Windows.Forms.FileDialogExtender
 
 		internal static uint OriginalDlgWidth
 		{
-			get { return FileDialogControlBase._originalDlgWidth; }
-			set { FileDialogControlBase._originalDlgWidth = value; }
+			get { return _originalDlgWidth; }
+			set { _originalDlgWidth = value; }
 		}
 
 		internal static uint OriginalDlgHeight
 		{
-			get { return FileDialogControlBase._originalDlgHeight; }
-			set { FileDialogControlBase._originalDlgHeight = value; }
+			get { return _originalDlgHeight; }
+			set { _originalDlgHeight = value; }
 		}
 
 		[Browsable(false)]
@@ -131,42 +127,26 @@ namespace SIL.Windows.Forms.FileDialogExtender
 				_StartLocation = value;
 				if (DesignMode)
 				{
-					this.Refresh();
+					Refresh();
 				}
 			}
 		}
 
-		Size _OriginalCtrlSize;
-		internal Size OriginalCtrlSize
-		{
-			get { return _OriginalCtrlSize; }
-			set
-			{
-				_OriginalCtrlSize = value;
-			}
-		}
+		internal Size OriginalCtrlSize { get; set; }
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue(FolderViewMode.Default)]
-		public FolderViewMode FileDlgDefaultViewMode
-		{
-			get { return _DefaultViewMode; }
-			set { _DefaultViewMode = value; }
-		}
+		public FolderViewMode FileDlgDefaultViewMode { get; set; } = FolderViewMode.Default;
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue(FileDialogType.OpenFileDlg)]
-		public FileDialogType FileDlgType
-		{
-			get { return _FileDlgType; }
-			set { _FileDlgType = value; }
-		}
+		public FileDialogType FileDlgType { get; set; }
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue("")]
 		public string FileDlgInitialDirectory
 		{
-			get { return DesignMode ? _InitialDirectory : MSDialog.InitialDirectory; }
+			get => DesignMode ? _InitialDirectory : MSDialog.InitialDirectory;
 			set
 			{
 				_InitialDirectory = value;
@@ -179,63 +159,55 @@ namespace SIL.Windows.Forms.FileDialogExtender
 		[DefaultValue("")]
 		public string FileDlgFileName
 		{
-			get { return DesignMode ? _FileName : MSDialog.FileName; }
-			set { _FileName = value; }
+			get => DesignMode ? _FileName : MSDialog.FileName;
+			set => _FileName = value;
 		}
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue("")]
-		public string FileDlgCaption
-		{
-			get { return _Caption; }
-			set { _Caption = value; }
-		}
+		public string FileDlgCaption { get; set; } = "Save";
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue("&Open")]
-		public string FileDlgOkCaption
-		{
-			get { return _OKCaption; }
-			set { _OKCaption = value; }
-		}
+		public string FileDlgOkCaption { get; set; } = "&Open";
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue("jpg")]
 		public string FileDlgDefaultExt
 		{
-			get { return DesignMode ? _DefaultExt : MSDialog.DefaultExt; }
-			set { _DefaultExt = value; }
+			get => DesignMode ? _DefaultExt : MSDialog.DefaultExt;
+			set => _DefaultExt = value;
 		}
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue("All files (*.*)|*.*")]
 		public string FileDlgFilter
 		{
-			get { return DesignMode ? _Filter : MSDialog.Filter; }
-			set { _Filter = value; }
+			get => DesignMode ? _Filter : MSDialog.Filter;
+			set => _Filter = value;
 		}
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue(1)]
 		public int FileDlgFilterIndex
 		{
-			get { return DesignMode ? _FilterIndex : MSDialog.FilterIndex; }
-			set { _FilterIndex = value; }
+			get => DesignMode ? _FilterIndex : MSDialog.FilterIndex;
+			set => _FilterIndex = value;
 		}
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue(true)]
 		public bool FileDlgAddExtension
 		{
-			get { return DesignMode ? _AddExtension : MSDialog.AddExtension; }
-			set { _AddExtension = value; }
+			get => DesignMode ? _AddExtension : MSDialog.AddExtension;
+			set => _AddExtension = value;
 		}
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue(true)]
 		public bool FileDlgEnableOkBtn
 		{
-			get { return _EnableOkBtn; }
+			get => _EnableOkBtn;
 			set
 			{
 				_EnableOkBtn = value;
@@ -248,25 +220,24 @@ namespace SIL.Windows.Forms.FileDialogExtender
 		[DefaultValue(true)]
 		public bool FileDlgCheckFileExists
 		{
-			get { return DesignMode ? _CheckFileExists : MSDialog.CheckFileExists; }
-			set
-			{ _CheckFileExists = value; }
+			get => DesignMode ? _CheckFileExists : MSDialog.CheckFileExists;
+			set => _CheckFileExists = value;
 		}
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue(false)]
 		public bool FileDlgShowHelp
 		{
-			get { return DesignMode ? _ShowHelp : MSDialog.ShowHelp; }
-			set { _ShowHelp = value; }
+			get => DesignMode ? _ShowHelp : MSDialog.ShowHelp;
+			set => _ShowHelp = value;
 		}
 
 		[Category("FileDialogExtenders")]
 		[DefaultValue(true)]
 		public bool FileDlgDereferenceLinks
 		{
-			get { return DesignMode ? _DereferenceLinks : MSDialog.DereferenceLinks; }
-			set { _DereferenceLinks = value; }
+			get => DesignMode ? _DereferenceLinks : MSDialog.DereferenceLinks;
+			set => _DereferenceLinks = value;
 		}
 		#endregion
 
@@ -280,13 +251,13 @@ namespace SIL.Windows.Forms.FileDialogExtender
 			{
 				if (MSDialog != null)
 				{
-					MSDialog.FileOk += new CancelEventHandler(FileDialogControlBase_ClosingDialog);
-					MSDialog.Disposed += new EventHandler(FileDialogControlBase_DialogDisposed);
-					MSDialog.HelpRequest += new EventHandler(FileDialogControlBase_HelpRequest);
+					MSDialog.FileOk += FileDialogControlBase_ClosingDialog;
+					MSDialog.Disposed += FileDialogControlBase_DialogDisposed;
+					MSDialog.HelpRequest += FileDialogControlBase_HelpRequest;
 					FileDlgEnableOkBtn = _EnableOkBtn;//that's designed time value
-					NativeMethods.SetWindowText(new HandleRef(_dlgWrapper,_dlgWrapper.Handle), _Caption);
+					NativeMethods.SetWindowText(new HandleRef(_dlgWrapper,_dlgWrapper.Handle), FileDlgCaption);
 					//will work only for open dialog, save dialog will be overriden internally by windows
-					NativeMethods.SetWindowText(new HandleRef(this,_hOKButton), _OKCaption);//SetDlgItemText fails too
+					NativeMethods.SetWindowText(new HandleRef(this,_hOKButton), FileDlgOkCaption);//SetDlgItemText fails too
 					//bool res = NativeMethods.SetDlgItemText(NativeMethods.GetParent(Handle), (int)ControlsId.ButtonOk, FileDlgOkCaption);
 				}
 			}
@@ -302,9 +273,9 @@ namespace SIL.Windows.Forms.FileDialogExtender
 				return;
 			if (MSDialog != null)
 			{
-				MSDialog.FileOk -= new CancelEventHandler(FileDialogControlBase_ClosingDialog);
-				MSDialog.Disposed -= new EventHandler(FileDialogControlBase_DialogDisposed);
-				MSDialog.HelpRequest -= new EventHandler(FileDialogControlBase_HelpRequest);
+				MSDialog.FileOk -= FileDialogControlBase_ClosingDialog;
+				MSDialog.Disposed -= FileDialogControlBase_DialogDisposed;
+				MSDialog.HelpRequest -= FileDialogControlBase_HelpRequest;
 				MSDialog.Dispose();
 				MSDialog = null;
 			}
@@ -351,21 +322,21 @@ namespace SIL.Windows.Forms.FileDialogExtender
 					Pen p = null;
 					try
 					{
-						switch (this.FileDlgStartLocation)
+						switch (FileDlgStartLocation)
 						{
 							case AddonWindowLocation.Right:
-								hb = new System.Drawing.Drawing2D.HatchBrush(HatchStyle.NarrowHorizontal, Color.Black, Color.Red);
+								hb = new HatchBrush(HatchStyle.NarrowHorizontal, Color.Black, Color.Red);
 								p = new Pen(hb, 5);
-								gr.DrawLine(p, 0, 0, 0, this.Height);
+								gr.DrawLine(p, 0, 0, 0, Height);
 								break;
 							case AddonWindowLocation.Bottom:
-								hb = new System.Drawing.Drawing2D.HatchBrush(HatchStyle.NarrowVertical, Color.Black, Color.Red);
+								hb = new HatchBrush(HatchStyle.NarrowVertical, Color.Black, Color.Red);
 								p = new Pen(hb, 5);
-								gr.DrawLine(p, 0, 0, this.Width, 0);
+								gr.DrawLine(p, 0, 0, Width, 0);
 								break;
 							case AddonWindowLocation.BottomRight:
 							default:
-								hb = new System.Drawing.Drawing2D.HatchBrush(HatchStyle.Sphere, Color.Black, Color.Red);
+								hb = new HatchBrush(HatchStyle.Sphere, Color.Black, Color.Red);
 								p = new Pen(hb, 5);
 								gr.DrawLine(p, 0, 0, 4, 4);
 								break;
@@ -373,10 +344,8 @@ namespace SIL.Windows.Forms.FileDialogExtender
 					}
 					finally
 					{
-						if (p != null)
-							p.Dispose();
-						if (hb != null)
-							hb.Dispose();
+						p?.Dispose();
+						hb?.Dispose();
 					}
 				}
 			}
@@ -412,15 +381,15 @@ namespace SIL.Windows.Forms.FileDialogExtender
 		public DialogResult ShowDialog(IWin32Window owner)
 		{
 			DialogResult returnDialogResult = DialogResult.Cancel;
-			if (this.IsDisposed)
+			if (IsDisposed)
 				return returnDialogResult;
 			if (owner == null || owner.Handle == IntPtr.Zero)
 			{
 				WindowWrapper wr = new WindowWrapper(System.Diagnostics.Process.GetCurrentProcess().MainWindowHandle);
 				owner = wr;
 			}
-			OriginalCtrlSize = this.Size;
-			_MSdialog = (FileDlgType == FileDialogType.OpenFileDlg)?new OpenFileDialog()as FileDialog:new SaveFileDialog() as FileDialog;
+			OriginalCtrlSize = Size;
+			_MSdialog = FileDlgType == FileDialogType.OpenFileDlg ? new OpenFileDialog() : new SaveFileDialog() as FileDialog;
 			_dlgWrapper = new WholeDialogWrapper(this);
 			OnPrepareMSDialog();
 			if (!_hasRunInitMSDialog)
@@ -432,8 +401,8 @@ namespace SIL.Windows.Forms.FileDialogExtender
 					AutoUpgradeInfo.SetValue(MSDialog, false, null);
 				returnDialogResult = _MSdialog.ShowDialog(owner);
 			}
-			// Sometimes if you open a animated .gif on the preview and the Form is closed, .Net class throw an exception
-			// Lets ignore this exception and keep closing the form.
+			// Sometimes if you open an animated .gif on the preview and the Form is closed, an
+			// exception is thrown. Let's ignore this exception and keep closing the form.
 			catch (ObjectDisposedException)
 			{
 			}
@@ -447,26 +416,26 @@ namespace SIL.Windows.Forms.FileDialogExtender
 		internal DialogResult ShowDialogExt(FileDialog fdlg,IWin32Window owner)
 		{
 			DialogResult returnDialogResult = DialogResult.Cancel;
-			if (this.IsDisposed)
+			if (IsDisposed)
 				return returnDialogResult;
 			if (owner == null || owner.Handle == IntPtr.Zero)
 			{
 				WindowWrapper wr = new WindowWrapper(System.Diagnostics.Process.GetCurrentProcess().MainWindowHandle);
 				owner = wr;
 			}
-			OriginalCtrlSize = this.Size;
+			OriginalCtrlSize = Size;
 			MSDialog = fdlg;
 			_dlgWrapper = new WholeDialogWrapper(this);
 
 			try
 			{
-				System.Reflection.PropertyInfo AutoUpgradeInfo = MSDialog.GetType().GetProperty("AutoUpgradeEnabled");
-				if (AutoUpgradeInfo != null)
-					AutoUpgradeInfo.SetValue(MSDialog, false, null);
+				var autoUpgradeInfo = MSDialog.GetType().GetProperty("AutoUpgradeEnabled");
+				if (autoUpgradeInfo != null)
+					autoUpgradeInfo.SetValue(MSDialog, false, null);
 				returnDialogResult = _MSdialog.ShowDialog(owner);
 			}
-			// Sometimes if you open a animated .gif on the preview and the Form is closed, .Net class throw an exception
-			// Lets ignore this exception and keep closing the form.
+			// Sometimes if you open an animated .gif on the preview and the Form is closed, an
+			// exception is thrown. Let's ignore this exception and keep closing the form.
 			catch (ObjectDisposedException)
 			{
 			}
@@ -503,7 +472,7 @@ namespace SIL.Windows.Forms.FileDialogExtender
 		#endregion
 		#region helper types
 
-		public class WindowWrapper : System.Windows.Forms.IWin32Window
+		public class WindowWrapper : IWin32Window
 		{
 			public WindowWrapper(IntPtr handle)
 			{